Extract schema parameters
